refactor: rename mod merge_tree to partition_tree
