+++ src/mito2/src/memtable/partition_tree/partition.rs
@@ -28,14 +28,14 @@ use store_api::storage::ColumnId;
 
 use crate::error::Result;
 use crate::memtable::key_values::KeyValue;
-use crate::memtable::merge_tree::data::{DataBatch, DataParts, DATA_INIT_CAP};
-use crate::memtable::merge_tree::dedup::DedupReader;
-use crate::memtable::merge_tree::metrics::WriteMetrics;
-use crate::memtable::merge_tree::shard::{
+use crate::memtable::partition_tree::data::{DataBatch, DataParts, DATA_INIT_CAP};
+use crate::memtable::partition_tree::dedup::DedupReader;
+use crate::memtable::partition_tree::metrics::WriteMetrics;
+use crate::memtable::partition_tree::shard::{
     BoxedDataBatchSource, Shard, ShardMerger, ShardNode, ShardSource,
 };
-use crate::memtable::merge_tree::shard_builder::ShardBuilder;
-use crate::memtable::merge_tree::{MergeTreeConfig, PkId};
+use crate::memtable::partition_tree::shard_builder::ShardBuilder;
+use crate::memtable::partition_tree::{MergeTreeConfig, PkId};
 use crate::metrics::MERGE_TREE_READ_STAGE_ELAPSED;
 use crate::read::{Batch, BatchBuilder};
 use crate::row_converter::{McmpRowCodec, RowCodec};
