+++ src/mito2/src/metrics.rs
@@ -278,23 +278,23 @@ lazy_static! {
         .with_label_values(&["flush", "intermediate"]);
     // ------- End of index metrics.
 
-    /// Merge tree memtable data buffer freeze metrics
+    /// Partition tree memtable data buffer freeze metrics
     pub static ref MERGE_TREE_DATA_BUFFER_FREEZE_STAGE_ELAPSED: HistogramVec = register_histogram_vec!(
-        "greptime_merge_tree_buffer_freeze_stage_elapsed",
-        "mito merge tree data buffer freeze stage elapsed",
+        "greptime_partition_tree_buffer_freeze_stage_elapsed",
+        "mito partition tree data buffer freeze stage elapsed",
         &[STAGE_LABEL],
         vec![0.005, 0.01, 0.05, 0.1, 0.5, 1.0, 5.0, 10.0, 60.0]
     )
     .unwrap();
 
-    /// Merge tree memtable read path metrics
+    /// Partition tree memtable read path metrics
     pub static ref MERGE_TREE_READ_STAGE_ELAPSED: HistogramVec = register_histogram_vec!(
-        "greptime_merge_tree_read_stage_elapsed",
-        "mito merge tree read stage elapsed",
+        "greptime_partition_tree_read_stage_elapsed",
+        "mito partition tree read stage elapsed",
         &[STAGE_LABEL],
         vec![0.005, 0.01, 0.05, 0.1, 0.5, 1.0, 5.0, 10.0, 60.0]
     )
     .unwrap();
 
-    // ------- End of merge tree memtable metrics.
+    // ------- End of partition tree memtable metrics.
 }
